requirements.txt
# 核心依赖
python-dotenv>=1.0.0

# 数据库
psycopg2-binary>=2.9.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# LangChain 核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# 向量存储
faiss-cpu>=1.7.4

# 数据验证
pydantic>=2.0.0

# 生物医学工具
biopython>=1.80

# 机器学习/嵌入
sentence-transformers>=2.2.0
transformers>=4.30.0

# HTTP 请求（如果需要）
requests>=2.31.0
